By default, use 127.0.0.1:8500 as public host + port
Only us can successfuly use test.regovar.org, whereas 127.0.0.1 will work
for anybody using a local deployment.

The real fix for issues with this is of course to properly set this value.
A commit for this in the SaltStack states will follow. The Docker install should
already handle this properly.